Delete election analysis.txt
Deleting this file because initially when the starter code in the challenge had this code and the output was generated in this file
# Add a variable to save the file to a path.
file_to_save = os.path.join("analysis", "election_analysis.txt")
But in the instructions for final submissions, it mentions the output file name as "election_results.txt". So re-generated the output in election_results.txt".